PLT-135: Refactor API continued. break down Api.Common module
