+++ plutus-core/testlib/PlutusCore/Test.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE FlexibleInstances      #-}
 {-# LANGUAGE FunctionalDependencies #-}
-{-# LANGUAGE OverloadedStrings      #-}
 {-# LANGUAGE RankNTypes             #-}
 {-# LANGUAGE TypeApplications       #-}
 {-# LANGUAGE UndecidableInstances   #-}
@@ -61,6 +60,7 @@ import Control.Lens
 import Control.Monad.Except
 import Control.Monad.Reader
 import Control.Monad.State
+import Data.Either.Extras
 import Data.Text (Text)
 import Hedgehog
 import Prettyprinter qualified as PP
@@ -140,7 +140,7 @@ catchAll :: a -> ExceptT SomeException IO a
 catchAll value = ExceptT $ try @SomeException (evaluate value)
 
 rethrow :: ExceptT SomeException IO a -> IO a
-rethrow = fmap (either throw id) . runExceptT
+rethrow = fmap unsafeFromEither . runExceptT
 
 runTPlc
     :: ToTPlc a TPLC.DefaultUni TPLC.DefaultFun
@@ -171,7 +171,7 @@ runUPlcProfile values = do
     ps <- traverse toUPlc values
     let (UPLC.Program _ _ t) = foldl1 UPLC.applyProgram ps
         (result, logOut) = UPLC.evaluateCek UPLC.logEmitter TPLC.defaultCekParameters t
-    res <- either (throwError . SomeException) pure result
+    res <- fromRightM (throwError . SomeException) result
     pure (res, logOut)
 
 -- For the profiling executable.
@@ -185,7 +185,7 @@ runUPlcProfileExec values = do
     ps <- traverse toUPlc values
     let (UPLC.Program _ _ t) = foldl1 UPLC.applyProgram ps
         (result, logOut) = UPLC.evaluateCek UPLC.logWithTimeEmitter TPLC.defaultCekParameters t
-    res <- either (throwError . SomeException) pure result
+    res <- fromRightM (throwError . SomeException) result
     pure (res, logOut)
 
 ppCatch :: PrettyPlc a => ExceptT SomeException IO a -> IO (Doc ann)
