+++ plutus-core/untyped-plutus-core/test/Evaluation/Builtins/Coherence.hs
@@ -26,7 +26,7 @@ typeSchemeToRuntimeScheme (TypeSchemeAll _ schK) =
 test_TypeSchemesAndRuntimeSchemesAgree :: TestTree
 test_TypeSchemesAndRuntimeSchemesAgree =
     testCase "type schemes are coherent with runtime schemes" $
-        for_ (enumeration @DefaultFun) $ \fun ->
+        for_ (enumerate @DefaultFun) $ \fun ->
             case toBuiltinMeaning @_ @_ @(Term TyName Name DefaultUni DefaultFun ()) fun of
                 BuiltinMeaning typeSch _ (BuiltinRuntimeOptions runtimeSch _ _ _) ->
                     typeSchemeToRuntimeScheme typeSch @?= runtimeSch
