--- plutus-ledger-api/src/Data/Either/Extras.hs
@@ -1,7 +1,0 @@
-module Data.Either.Extras
-  ( unsafeFromEither
-  ) where
-
-unsafeFromEither :: Either String a -> a
-unsafeFromEither (Left err)    = error err
-unsafeFromEither (Right value) = value
