+++ plutus-ledger-api/src/PlutusLedgerApi/V1/Bytes.hs
@@ -1,24 +1,19 @@
-{-# LANGUAGE DataKinds                  #-}
-{-# LANGUAGE DeriveAnyClass             #-}
-{-# LANGUAGE DeriveGeneric              #-}
-{-# LANGUAGE DerivingStrategies         #-}
-{-# LANGUAGE DerivingVia                #-}
-{-# LANGUAGE FlexibleContexts           #-}
-{-# LANGUAGE FlexibleInstances          #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
-{-# LANGUAGE OverloadedStrings          #-}
-{-# LANGUAGE TemplateHaskell            #-}
-{-# LANGUAGE TypeApplications           #-}
+{-# LANGUAGE DeriveAnyClass  #-}
+{-# LANGUAGE DerivingVia     #-}
+{-# LANGUAGE TemplateHaskell #-}
+
 {-# OPTIONS_GHC -Wno-orphans            #-}
 
-module PlutusLedgerApi.V1.Bytes ( LedgerBytes (..)
-                , fromHex
-                , bytes
-                , fromBytes
-                , encodeByteString
-                ) where
+module PlutusLedgerApi.V1.Bytes
+    ( LedgerBytes (..)
+    , fromHex
+    , bytes
+    , fromBytes
+    , encodeByteString
+    ) where
 
 import Control.DeepSeq (NFData)
+import Control.Exception
 import Data.ByteString qualified as BS
 import Data.ByteString.Base16 qualified as Base16
 import Data.ByteString.Internal (c2w, w2c)
@@ -28,39 +23,44 @@ import Data.Text qualified as Text
 import Data.Text.Encoding qualified as TE
 import Data.Word (Word8)
 import GHC.Generics (Generic)
-import PlutusTx qualified as PlutusTx
-import PlutusTx.Lift
+import PlutusTx
 import PlutusTx.Prelude qualified as P
 import Prettyprinter.Extras (Pretty, PrettyShow (..))
 
-fromHex :: BS.ByteString -> Either String LedgerBytes
+data LedgerBytesError =
+    UnpairedDigit
+    | NotHexit Char
+    deriving stock (Show)
+    deriving anyclass (Exception)
+
+fromHex :: BS.ByteString -> Either LedgerBytesError LedgerBytes
 fromHex = fmap (LedgerBytes . P.toBuiltin) . asBSLiteral
     where
 
-    handleChar :: Word8 -> Either String Word8
+    handleChar :: Word8 -> Either LedgerBytesError Word8
     handleChar x
         | x >= c2w '0' && x <= c2w '9' = Right (x - c2w '0') -- hexits 0-9
         | x >= c2w 'a' && x <= c2w 'f' = Right (x - c2w 'a' + 10) -- hexits a-f
         | x >= c2w 'A' && x <= c2w 'F' = Right (x - c2w 'A' + 10) -- hexits A-F
-        | otherwise = Left ("not a hexit: " <> show (w2c x) <> "")
+        | otherwise = Left $ NotHexit (w2c x)
 
     -- turns a pair of bytes such as "a6" into a single Word8
-    handlePair :: Word8 -> Word8 -> Either String Word8
+    handlePair :: Word8 -> Word8 -> Either LedgerBytesError Word8
     handlePair c c' = do
       n <- handleChar c
       n' <- handleChar c'
       pure $ (16 * n) + n'
 
-    asBytes :: [Word8] -> Either String [Word8]
+    asBytes :: [Word8] -> Either LedgerBytesError [Word8]
     asBytes []        = Right mempty
     asBytes (c:c':cs) = (:) <$> handlePair c c' <*> asBytes cs
-    asBytes _         = Left "unpaired digit"
+    asBytes _         = Left UnpairedDigit
 
     -- parses a bytestring such as @a6b4@ into an actual bytestring
-    asBSLiteral :: BS.ByteString -> Either String BS.ByteString
+    asBSLiteral :: BS.ByteString -> Either LedgerBytesError BS.ByteString
     asBSLiteral = withBytes asBytes
         where
-          withBytes :: ([Word8] -> Either String [Word8]) -> BS.ByteString -> Either String BS.ByteString
+          withBytes :: ([Word8] -> Either LedgerBytesError [Word8]) -> BS.ByteString -> Either LedgerBytesError BS.ByteString
           withBytes f = fmap BS.pack . f . BS.unpack
 
 newtype LedgerBytes = LedgerBytes { getLedgerBytes :: P.BuiltinByteString }
