+++ plutus-ledger-api/src/PlutusLedgerApi/V1/Credential.hs
@@ -1,19 +1,14 @@
-{-# LANGUAGE DataKinds         #-}
 {-# LANGUAGE DeriveAnyClass    #-}
-{-# LANGUAGE DerivingVia       #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell   #-}
+
 {-# OPTIONS_GHC -fno-specialise #-}
 {-# OPTIONS_GHC -Wno-simplifiable-class-constraints #-}
 {-# OPTIONS_GHC -fno-omit-interface-pragmas #-}
 
-{-
-
-Address and staking address credentials for outputs.
-
--}
-module PlutusLedgerApi.V1.Credential(
-    StakingCredential(..)
+-- |Address and staking address credentials for outputs.
+module PlutusLedgerApi.V1.Credential
+    ( StakingCredential(..)
     , Credential(..)
     ) where
