+++ plutus-ledger-api/src/PlutusLedgerApi/V1/Tx.hs
@@ -1,43 +1,41 @@
 {-# LANGUAGE DeriveAnyClass    #-}
-{-# LANGUAGE DeriveGeneric     #-}
 {-# LANGUAGE DerivingVia       #-}
-{-# LANGUAGE FlexibleContexts  #-}
-{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards   #-}
 {-# LANGUAGE TemplateHaskell   #-}
+
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-# OPTIONS_GHC -fno-specialise #-}
 {-# OPTIONS_GHC -fno-omit-interface-pragmas #-}
 
-module PlutusLedgerApi.V1.Tx(
+module PlutusLedgerApi.V1.Tx
+    (
     -- * Transactions
-    TxId (..),
-    ScriptTag (..),
-    RedeemerPtr (..),
-    Redeemers,
+      TxId (..)
+    , ScriptTag (..)
+    , RedeemerPtr (..)
+    , Redeemers
     -- * Transaction outputs
-    TxOut(..),
-    TxOutRef(..),
-    isPubKeyOut,
-    isPayToScriptOut,
-    outAddress,
-    outValue,
-    txOutPubKey,
-    txOutDatum,
-    pubKeyHashTxOut,
+    , TxOut(..)
+    , TxOutRef(..)
+    , isPubKeyOut
+    , isPayToScriptOut
+    , outAddress
+    , outValue
+    , txOutPubKey
+    ,txOutDatum
+    ,pubKeyHashTxOut
     -- * Transaction inputs
-    TxInType(..),
-    TxIn(..),
-    inRef,
-    inType,
-    inScripts,
-    pubKeyTxIn,
-    scriptTxIn,
-    pubKeyTxIns,
-    scriptTxIns,
+    , TxInType(..)
+    , TxIn(..)
+    , inRef
+    , inType
+    , inScripts
+    , pubKeyTxIn
+    , scriptTxIn
+    , pubKeyTxIns
+    , scriptTxIns
     ) where
 
 import Control.DeepSeq (NFData)
