+++ plutus-ledger-api/test/Spec/CostModelParams.hs
@@ -1,14 +1,14 @@
 {-# LANGUAGE TypeApplications #-}
 module Spec.CostModelParams where
 
-import PlutusLedgerApi.V1.EvaluationContext as V1
-import PlutusLedgerApi.V2.EvaluationContext as V2
+import PlutusLedgerApi.Common
+import PlutusLedgerApi.V1 as V1
+import PlutusLedgerApi.V2 as V2
 
 import PlutusCore.Evaluation.Machine.BuiltinCostModel as Plutus
 import PlutusCore.Evaluation.Machine.CostModelInterface as Plutus
 import PlutusCore.Evaluation.Machine.ExBudgetingDefaults as Plutus
 import PlutusCore.Evaluation.Machine.MachineParameters as Plutus
-
 
 import Barbies
 import Control.Lens
@@ -16,6 +16,7 @@ import Data.Either
 import Data.List.Extra
 import Data.Map as Map
 import Data.Maybe
+import Data.Set as Set
 import Data.Text qualified as Text
 import Test.Tasty
 import Test.Tasty.HUnit
@@ -26,22 +27,27 @@ tests =
     "costModelParams"
     [ testCase "length" $ do
             166 @=? length (enumerate @V1.ParamName)
-            166 @=? length V1.costModelParamNames
+            166 @=? length v1CostModelParamNames
             175 @=? length (enumerate @V2.ParamName)
-            175 @=? length V2.costModelParamNames
+            175 @=? length v2CostModelParamNames
     , testCase "text" $ do
             -- this depends on the fact that V1/V2 are alphabetically-ordered; does not have to hold for future versions
-            altV1CostModelParamNames @=? V1.costModelParamNames
+            altV1CostModelParamNames @=? v1CostModelParamNames
             -- this depends on the fact that V1/V2 are alphabetically-ordered; does not have to hold for future versions
-            Map.keys (fromJust Plutus.defaultCostModelParams) @=? V2.costModelParamNames
+            Map.keys (fromJust Plutus.defaultCostModelParams) @=? v2CostModelParamNames
     , testCase "context length" $ do
             let defaultCostValues = Map.elems $ fromJust defaultCostModelParams
             -- the defaultcostmodelparams reflects only the latest version V2, so this should succeed because the lengths match
             assertBool "wrong number of arguments in V2.mkContext" $ isRight $ V2.mkEvaluationContext defaultCostValues
             -- this one should not succeed because V1 evaluation-context expects less number of arguments.
             assertBool "wrong number of arguments in V1.mkContext" $ isLeft $ V1.mkEvaluationContext defaultCostValues
+    , testCase "cost model parameters" $
+         -- v1 is missing some cost model parameters because new builtins are added in v2
+         assertBool "v1 params is proper subset of v2 params" $ Set.fromList v1CostModelParamNames `Set.isProperSubsetOf` Set.fromList v2CostModelParamNames
     ]
 
+v1CostModelParamNames :: [Text.Text]
+v1CostModelParamNames = Text.pack . showParamName <$> enumerate @V1.ParamName
 
 -- | An alternative, older implementation of calculating V1's costmodelparamnames.
 altV1CostModelParamNames :: [Text.Text]
@@ -63,3 +69,7 @@ altV1CostModelParamNames = Map.keys $ fromJust $ extractCostModelParams $
             , paramVerifyEcdsaSecp256k1Signature = mempty
             , paramVerifySchnorrSecp256k1Signature = mempty
             }
+
+v2CostModelParamNames :: [Text.Text]
+v2CostModelParamNames = Text.pack . showParamName <$> enumerate @V2.ParamName
+
