+++ plutus-ledger-api/test/Spec/Eval.hs
@@ -125,15 +125,15 @@ testScripts = "v1-scripts" `testWith` evalScripts
 
 
 {-| Evaluates scripts as they will be evaluated on-chain, by using the evaluation function we provide for the ledger.
-Notably, this goes via serializing and deserializing the program, so we can see any errors that might arise from that.
+Notably, this goes via serialising and deserialising the program, so we can see any errors that might arise from that.
 -}
 testAPI :: TestTree
 testAPI = "v1-api" `testWith` evalAPI vasilPV
 
 evalAPI :: ProtocolVersion -> UPLC.Term DeBruijn DefaultUni DefaultFun () -> Bool
 evalAPI pv t =
-    -- handcraft a serialized script
-    let s :: SerializedScript = BSS.toShort . BSL.toStrict . CBOR.serialise $ Script $ mkProg t
+    -- handcraft a serialised script
+    let s :: SerialisedScript = BSS.toShort . BSL.toStrict . CBOR.serialise $ Script $ mkProg t
     in isRight $ snd $ Api.evaluateScriptRestricting pv Quiet evalCtxForTesting (unExRestrictingBudget enormousBudget) s []
 
 -- Test a given eval function against the expected results.
