+++ plutus-ledger-api/testlib/PlutusLedgerApi/Test/EvaluationContext.hs
@@ -7,7 +7,7 @@ import PlutusCore.Evaluation.Machine.CostModelInterface as Plutus
 import PlutusCore.Evaluation.Machine.ExBudgetingDefaults qualified as Plutus
 import PlutusLedgerApi.Common
 
-import Control.Exception
+import Data.Either.Extras
 import Data.Maybe
 
 -- | The raw cost model params, only to be used for testing purposes.
@@ -16,4 +16,4 @@ costModelParamsForTesting = fromJust Plutus.defaultCostModelParams
 
 -- | only to be for testing purposes: make an evaluation context by applying an empty set of protocol parameters
 evalCtxForTesting :: EvaluationContext
-evalCtxForTesting = either throw id $ mkDynEvaluationContext costModelParamsForTesting
+evalCtxForTesting = unsafeFromEither $ mkDynEvaluationContext costModelParamsForTesting
