+++ plutus-ledger-api/testlib/PlutusLedgerApi/Test/Examples.hs
@@ -1,7 +1,5 @@
 {-# LANGUAGE TypeApplications #-}
-{-|
-This module contains example values to be used for testing. These should NOT be used in non-test code!
--}
+-- | This module contains example values to be used for testing. These should NOT be used in non-test code!
 module PlutusLedgerApi.Test.Examples (alwaysSucceedingNAryFunction, alwaysFailingNAryFunction, summingFunction, saltFunction) where
 
 import PlutusCore qualified as PLC
@@ -23,7 +21,7 @@ It seems better therefore to avoid depending on Plutus Tx in any "core" projects
 -}
 
 -- | Creates a script which has N arguments, and always succeeds.
-alwaysSucceedingNAryFunction :: Natural -> SerializedScript
+alwaysSucceedingNAryFunction :: Natural -> SerialisedScript
 alwaysSucceedingNAryFunction n = toShort $ toStrict $ serialise $ Scripts.Script $ UPLC.Program () (PLC.defaultVersion ()) (body n)
     where
         -- No more arguments! The body can be anything that doesn't fail, so we return `\x . x`
@@ -32,7 +30,7 @@ alwaysSucceedingNAryFunction n = toShort $ toStrict $ serialise $ Scripts.Script
         body i = UPLC.LamAbs () (UPLC.DeBruijn 0) $ body (i-1)
 
 -- | Creates a script which has N arguments, and always fails.
-alwaysFailingNAryFunction :: Natural -> SerializedScript
+alwaysFailingNAryFunction :: Natural -> SerialisedScript
 alwaysFailingNAryFunction n = toShort $ toStrict $ serialise $ Scripts.Script $ UPLC.Program () (PLC.defaultVersion ()) (body n)
     where
         -- No more arguments! The body should be error.
@@ -40,13 +38,13 @@ alwaysFailingNAryFunction n = toShort $ toStrict $ serialise $ Scripts.Script $ 
         -- We're using de Bruijn indices, so we can use the same binder each time!
         body i = UPLC.LamAbs () (UPLC.DeBruijn 0) $ body (i-1)
 
-summingFunction :: SerializedScript
+summingFunction :: SerialisedScript
 summingFunction = toShort $ toStrict $ serialise $ Scripts.Script $ UPLC.Program () (PLC.defaultVersion ()) body
     where
         body = UPLC.Apply () (UPLC.Apply () (UPLC.Builtin () PLC.AddInteger) (PLC.mkConstant @Integer () 1)) (PLC.mkConstant @Integer () 2)
 
 -- | Wrap a script with lambda/app so that, for instance, it has a different hash but the same behavior.
-saltFunction :: Integer -> SerializedScript -> SerializedScript
+saltFunction :: Integer -> SerialisedScript -> SerialisedScript
 saltFunction salt b0 = toShort $ toStrict $ serialise $ Scripts.Script $ UPLC.Program () version body
     where
         Scripts.Script (UPLC.Program () version b1) = deserialise $ fromStrict $ fromShort b0
