+++ plutus-tx-plugin/test/Lib.hs
@@ -13,6 +13,7 @@ module Lib where
 import Control.Exception
 import Control.Lens
 import Control.Monad.Except
+import Data.Either.Extras
 import Data.Text (Text)
 import Flat (Flat)
 import Test.Tasty.Extras
@@ -43,7 +44,7 @@ runPlcCek :: ToUPlc a PLC.DefaultUni PLC.DefaultFun => [a] -> ExceptT SomeExcept
 runPlcCek values = do
      ps <- traverse toUPlc values
      let p = foldl1 UPLC.applyProgram ps
-     either (throwError . SomeException) pure $ evaluateCekNoEmit PLC.defaultCekParameters (p ^. UPLC.progTerm)
+     fromRightM (throwError . SomeException) $ evaluateCekNoEmit PLC.defaultCekParameters (p ^. UPLC.progTerm)
 
 runPlcCekTrace ::
      ToUPlc a PLC.DefaultUni PLC.DefaultFun =>
@@ -53,7 +54,7 @@ runPlcCekTrace values = do
      ps <- traverse toUPlc values
      let p = foldl1 UPLC.applyProgram ps
      let (result, TallyingSt tally _, logOut) = runCek PLC.defaultCekParameters tallying logEmitter (p ^. UPLC.progTerm)
-     res <- either (throwError . SomeException) pure result
+     res <- fromRightM (throwError . SomeException) result
      pure (logOut, tally, res)
 
 goldenEvalCek :: ToUPlc a PLC.DefaultUni PLC.DefaultFun => String -> [a] -> TestNested
